feat(backend): alias
Uses the new `fetch` to delegates the request to underlying remote.

While simple, this demonstrates our approach to backend that wraps other
backend.

For example, a `compress` backend can grab the response from underlying
backend and pipe through a `CompressionStream`.

Or a `crypt` backend that (de)crypt the response from underlying backend